Corrected problem of multiple selections of land uses for one triangle mesh
Multiple land use selections were assigned to the same mesh triangle.
Also, there were some land use selections that did not match the correct
land use in the maps. The problem was identified in the section when the
maximum land use areas were matched to their land use designation. Just
matching the value of the maximum sum land use area to the original set
with all the summation areas produced multiple matches of area sums in
different sections of the complete area summation database. Therefore it
selected multiple land uses for some triangles (the number of times
their max summation are matched other summation areas in the database).
The problem was solved by splitting the database into groups based on
each triangle in the mesh and selecting the maximum summation area
within each group, extract the information of the LU with the maximum
summation area within the group and regrouping the extracted information
into a data frame again.